feat: LocationView, LocationViewModel 분리
